기능 추가
Read

+++ src/main/java/kyj/rest/mvc/demo/repository/BoardRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import kyj.rest.mvc.demo.entity.PostEntity;
 
 @Repository
-public interface BoardRepository extends JpaRepository<PostEntity, Long>{
+public interface BoardRepository extends JpaRepository<PostEntity, Integer>{
 	
 }
